chore(registration-react): refactor move catalog field-tree-los to catalog form-los and update imports
