+++ applications/registration-react/src/pages/dataset-registration-page/form-los/field-tree-los/field-tree-los.scss
@@ -1,4 +1,4 @@
-@import '../../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
+@import '../../../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
 
 .tree-view {
   overflow-x: hidden;
